# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: English Anywhere Lab API
  version: 0.4.0
  description: |
    Single OpenAPI contract for English Anywhere Lab.
    Base path is /api/v1.
    This file is the only source of truth for API schema.

    Operations marked with `x-implemented: true` are live and covered by
    integration tests. Operations with `x-implemented: false` are roadmap
    placeholders — do NOT include them in contract tests or QA automation.
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
paths:
...
  /progress/weekly-report:
    get:
      tags: [Progress]
      summary: Get weekly report aggregation
      operationId: getWeeklyReport
      x-implemented: true
      parameters:
        - in: query
          name: week_start
          required: true
          schema:
            type: string
            format: date
          description: ISO 8601 date of a Monday (start of the week)
      responses:
        '200':
          description: Weekly report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeeklyReportResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /progress/monthly-report:
    get:
      tags: [Progress]
      summary: Get monthly report with skill breakdown and weakness analysis
      operationId: getMonthlyReport
      x-implemented: true
      parameters:
        - in: query
          name: month
          required: true
          schema:
            type: string
            pattern: '^\d{4}-(0[1-9]|1[0-2])$'
          description: Year-month in YYYY-MM format
      responses:
        '200':
          description: Monthly report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MonthlyReportResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
      required: [code, message]
...
    WeeklyReportDailyPoint:
      type: object
      properties:
        date:
          type: string
          format: date
        minutes_learned:
          type: integer
        lessons_completed:
          type: integer
        cards_new:
          type: integer
        cards_reviewed:
          type: integer
        review_accuracy:
          type: number
          format: float
          nullable: true
        listening_minutes:
          type: integer
        speaking_tasks:
          type: integer
        writing_tasks:
          type: integer
        streak_count:
          type: integer
      required: [date, minutes_learned, lessons_completed, cards_new, cards_reviewed, listening_minutes, speaking_tasks, writing_tasks, streak_count]
    ReviewHealth:
      type: object
      properties:
        again:
          type: integer
        hard:
          type: integer
        good:
          type: integer
        easy:
          type: integer
        total:
          type: integer
        accuracy:
          type: number
          format: float
          nullable: true
      required: [again, hard, good, easy, total]
    WeeklyComparison:
      type: object
      properties:
        minutes_delta:
          type: integer
        active_days_delta:
          type: integer
        cards_reviewed_delta:
          type: integer
        lessons_delta:
          type: integer
        accuracy_delta:
          type: number
          format: float
          nullable: true
      required: [minutes_delta, active_days_delta, cards_reviewed_delta, lessons_delta]
    WeeklyReportResponse:
      type: object
      properties:
        week_start:
          type: string
          format: date
        total_minutes:
          type: integer
        active_days:
          type: integer
        cards_reviewed:
          type: integer
        cards_new:
          type: integer
        lessons_completed:
          type: integer
        listening_minutes:
          type: integer
        speaking_tasks:
          type: integer
        writing_tasks:
          type: integer
        streak:
          type: integer
        weekly_goal_days:
          type: integer
        goal_achieved:
          type: boolean
        review_health:
          $ref: '#/components/schemas/ReviewHealth'
        daily_breakdown:
          type: array
          items:
            $ref: '#/components/schemas/WeeklyReportDailyPoint'
        previous_week_comparison:
          $ref: '#/components/schemas/WeeklyComparison'
          nullable: true
      required: [week_start, total_minutes, active_days, cards_reviewed, cards_new, lessons_completed, listening_minutes, speaking_tasks, writing_tasks, streak, weekly_goal_days, goal_achieved, review_health, daily_breakdown]
    SkillMetric:
      type: object
      properties:
        skill:
          type: string
          enum: [listening, speaking, writing, reading]
        value:
          type: integer
        percentage:
          type: number
          format: float
      required: [skill, value, percentage]
    SkillBreakdown:
      type: object
      properties:
        listening:
          $ref: '#/components/schemas/SkillMetric'
        speaking:
          $ref: '#/components/schemas/SkillMetric'
        writing:
          $ref: '#/components/schemas/SkillMetric'
        reading:
          $ref: '#/components/schemas/SkillMetric'
      required: [listening, speaking, writing, reading]
    WeaknessItem:
      type: object
      properties:
        skill:
          type: string
          enum: [listening, speaking, writing, reading]
        reason:
          type: string
          enum: [low_activity, declining, below_average]
        value:
          type: integer
        prev_value:
          type: integer
          nullable: true
      required: [skill, reason, value]
    MonthlyComparison:
      type: object
      properties:
        minutes_delta:
          type: integer
        active_days_delta:
          type: integer
        cards_reviewed_delta:
          type: integer
        lessons_delta:
          type: integer
        accuracy_delta:
          type: number
          format: float
          nullable: true
      required: [minutes_delta, active_days_delta, cards_reviewed_delta, lessons_delta]
    MonthlyReportResponse:
      type: object
      properties:
        month:
          type: string
          pattern: '^\d{4}-(0[1-9]|1[0-2])$'
        days_in_month:
          type: integer
        total_minutes:
          type: integer
        active_days:
          type: integer
        cards_reviewed:
          type: integer
        cards_new:
          type: integer
        lessons_completed:
          type: integer
        listening_minutes:
          type: integer
        speaking_tasks:
          type: integer
        writing_tasks:
          type: integer
        streak:
          type: integer
        monthly_goal_days:
          type: integer
        goal_achieved:
          type: boolean
        review_health:
          $ref: '#/components/schemas/ReviewHealth'
        daily_breakdown:
          type: array
          items:
            $ref: '#/components/schemas/WeeklyReportDailyPoint'
        skill_breakdown:
          $ref: '#/components/schemas/SkillBreakdown'
        weaknesses:
          type: array
          items:
            $ref: '#/components/schemas/WeaknessItem'
        previous_month_comparison:
          $ref: '#/components/schemas/MonthlyComparison'
          nullable: true
      required: [month, days_in_month, total_minutes, active_days, cards_reviewed, cards_new, lessons_completed, listening_minutes, speaking_tasks, writing_tasks, streak, monthly_goal_days, goal_achieved, review_health, daily_breakdown, skill_breakdown, weaknesses]
...